+++ src/filter-builder/filter-builder.d.ts
@@ -1,5 +1,6 @@
 import "../elements.js";
 
+export { applyFilters } from "../../types/filter-builder/apply-filters.js";
 export { RowanFilterBuilder } from "../../types/filter-builder/filter-builder.js";
 export type {
   RowanFilter,