+++ types/filter-builder/index.d.ts
@@ -1,1 +1,2 @@
+export { applyFilters } from "./apply-filters.js";
 export { RowanFilterBuilder } from "./filter-builder.js";